Separando as responsabilidades por pastas e criando a classe MainServices para guardar tudo que está padronizado nela
